vfs: fix ESTALE: mount ext3 partitions with 'data=journal'
Without this and without the lkl_sys_sync() in create_file, the
inode's metadata does not seem to be flushed to the disk.

When we create a new file the inode is allocated, an dentry is written
to the directory, the file's contents are written to the disk, but the
inode metadata is not.

When the file is accessed later, the on-disk inode has 0 in all fields
leading to a -ESTALE from file system calls (lkl_sys_stat,
lkl_sys_open, etc.)

I'm not sure why "data=journal" works ok, but it does.

Also, removing lkl_sys_sync() leads to -ESTALE too.

Signed-off-by: Lucian Adrian Grijincu <lucian.grijincu@gmail.com>